test(e2e): extract dual-controller suite into multi_controller_test.go
- Move the "Dual namespace-scoped controllers isolation" Context out of
  smoke_test.go into a new multi_controller_test.go labeled
  Label("multi-controller") so it can be selected by ginkgo label-filter.
- Repoint test/e2e/testdata/secondary-controller/kustomization.yaml to the
  new namespace-scoped/kubernetes overlay added in the prior PR; the
  --watch-namespace patch is dropped because the namespace-scoped component
  already injects it.
- Add make targets test-e2e-multi-controller and
  test-e2e-multi-controller-with-setup mirroring the smoke targets.

The patch-and-restore RBAC workaround stays in place for now; a later PR
removes it once the dedicated CI job runs the suite in a fresh kind
cluster.

Refs #1205
